requirements.txt
# Info-Getter Python 依赖

# 核心依赖
feedparser>=6.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pyyaml>=6.0.0

# Selenium 浏览器自动化
selenium>=4.15.0
webdriver-manager>=4.0.0

# 数据库
pymongo>=4.6.0

# 工具
python-dateutil>=2.8.0
lxml>=4.9.0
